block, bfq-mq: remove case of redirected bic from insert_request
Before commit "block, bfq-mq: postpone rq preparation to insert or
merge", the destination queue for a request was chosen by a different
hook than the one that then inserted the request. So, between the
execution of the two hooks, the bic of the process generating the
request could happen to be redirected to a different bfq_queue. As a
consequence, the destination bfq_queue stored in the request could be
wrong. Such an event does not need to ba handled any longer.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>