msm: kgsl: Don't allocate memory dynamically for temp command buffers
The temporary command buffer in _set_pagetable_gpu is only the size of a
single page, and _set_pagetable_gpu is never executed concurrently. It
is therefore easy to replace the dynamic command buffer allocation with
a static one to improve performance by avoiding the latency of dynamic
memory allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: reaPeR1010 <reaPeR10x10x@gmail.com>
Signed-off-by: Excalibur-99 <txexcalibur99@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>