msm: kgsl: Fix gdsc disable timeout during reset
Smmu driver takes a vote on CX gdsc before calling the kgsl pagefault
handler. If there is contention for device mutex in this path and the
dispatcher fault handler is holding this lock, trying to turn off CX
gdsc will fail during the reset. So to avoid blocking here, try to lock
device mutex and return if it fails.

Since we are modifying here, keep the pagefault header for UCHE block
short.

Change-Id: Idf1726960071d66b9112f6b4abdf19a49afacefd
Signed-off-by: Akhil P Oommen <akhilpo@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>