msm: kgsl: Don't wait for room in context queue when context is invalidated
In a condition where there is no room in the drawcontext queue and the
context is invalidated while sleeping in wait_event_interruptible_timeout,
we are now polling for room after wake up. This would lead to infinite
looping.

Hence, do not wait for room in context queue if context is invalidated.

Bug: 167124402
Change-Id: I12c80e71630da2f3316838bee37f13cfa3b8e7d9
Signed-off-by: Harshitha Sai Neelati <hsaine@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>