Makefile: clang-14: Disable 'unused-but-set-variable' for now
This warning is created when a variable is set but not used due to certain
reasons like it might be ifdef-ed with some condition or it might be a
part of a config or it be a debug config.

Anyway, we can hotfix by 'maybe_unused' macro but again we need to edit a
hell lot of drivers. Since this warning is harmless, we can just
suppress the warning as of now.
Anytime later in our free-time, we can just fix the warnings.

Fixes warnings:
../security/smack/smackfs.c:2879:16: warning: variable 'root_inode' set but not used [-Wunused-but-set-variable]
../block/blk-core.c:2270:7: warning: variable 'flags' set but not used [-Wunused-but-set-variable]
../mm/shmem.c:3281:27: warning: variable 'info' set but not used [-Wunused-but-set-variable]
../mm/vmscan.c:2689:17: warning: variable 'node_lru_pages' set but not used [-Wunused-but-set-variable]
../drivers/block/loop.c:674:16: warning: variable 'inode' set but not used [-Wunused-but-set-variable]
../mm/workingset.c:200:15: warning: variable 'nid' set but not used [-Wunused-but-set-variable]
..and many more

Change-Id: Idf8e03f7083d5fc9ff30081651bd4ac1f2dc5bd7
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>