zswap: break the strict dependency from lz4
From the beginning, the zswap block device has always required CRYPTO_LZ4 to be enabled,
since lz4 is hardcoded as the default compression algorithm.
As a consequence, when a different compression algorithm is hardcoded and the respective CRYPTO_<algo> is not enabled,
zswap would not be used therefore causing an issue when swap is needed.

This patch removes the hardcoded lz4 dependency and allows the user to select the default compression algorithm for zswap at build time.
The previous behaviour is kept, as the default compression algorithm is still lz4.

This patch was inspired by [1].

[1]: https://github.com/torvalds/linux/commit/3d711a382735d2c34d3ba2075a5aa83a894f4a57

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>