PM / suspend: Clear wakeups before running PM callbacks
PM callbacks can be used as an indicator that the system is suspending, and
as such, a wakeup may be issued outside the notifier in order to cancel an
ongoing suspend. Clearing wakeups after running the PM callbacks can thus
cause wakeups to be missed. Fix it by simply clearing wakeups prior to
running PM callbacks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I49b6f900284ff42b245be77548f6ca756cf585ee
Signed-off-by: Excalibur-99 <txexcalibur99@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>