devfreq/adreno_tz: make high refresh multiplier configurable
Meeting frame deadlines been shown to not have a linear response to frame rate.
There is no testing change between 1.5 and 1.25 here.

But let's make it configurable for more testing down the road and to make more
portable to other high refresh devices which do not reflect refresh rate in scaling.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Forenche <prahul2003@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>